Refactor transaction API tests and improve documentation
Moved tests from `src/api/transaction/mod.rs` to `src/api/transaction/tests.rs` to clean up the module file.
Updated documentation for `ReadOps` and `WriteOps` to include executable examples using the recommended closure-based API.
Added a comprehensive module-level example for transaction usage.
Fixed a syntax error in `src/index/vector/hnsw.rs` caused by a missing brace in `FilterCallbackGuard`.
